requirements.txt
huggingface-hub==0.25.1
numpy==2.1.1
opencv-python==4.10.0.84
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.2
pyarrow==17.0.0
pydeck==0.9.1
pydub==0.25.1
pywin32==305.1
PyYAML==6.0.2
safetensors==0.4.5
scipy==1.14.1
sounddevice==0.5.0
streamlit==1.38.0
toml==0.10.2
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
